fix non-exclusive pattern matches found by ghc 9.2
When getting zfs properties, if the output is not as expected, it's
better to skip it than it is to error out at runtime, probably.

When parsing a string to generate a PPA or AptSource, this causes
garbage to be generated when garbage is input. I'm not very happy about
this really, it seems it would be better to not have the IsString
instances for those, because IsString is best used when any string will
be accepted.

Sponsored-by: Brock Spratlen on Patreon